requirements.txt
requests
setuptools
pycryptodome